adjustment on errs messages

+++ src/useCases/DeleteUser/DeleteUserUseCase.ts
@@ -1,4 +1,3 @@
-import { DeleteUser } from "../../entities/DeleteUser";
 import { IUserRepository } from "../../repositories/IUserRepository";
 import { IDeleteUserRequestDTO } from "./DeleteUserUseCaseDTO";
 
@@ -6,6 +5,11 @@ export class DeleteUserUseCase {
   constructor(private userRepository: IUserRepository) {}
 
   async execute(data: IDeleteUserRequestDTO) {
+    const userExist = await this.userRepository.findById(data.id);
+    if (!userExist) {
+      throw new Error("User not found");
+    }
+
     await this.userRepository.delete(data.id);
   }
 }
